+++ lib/runtime-core/src/parse.rs
@@ -6,8 +6,8 @@ use crate::{
     backend::{CompilerConfig, RunnableModule},
     error::CompileError,
     module::{
-        DataInitializer, ExportIndex, ImportName, ModuleInfo, StringTable, StringTableBuilder,
-        TableInitializer,
+        DataInitializer, ExportIndex, ImportName, ModuleInfo, NameIndex, NamespaceIndex,
+        StringTable, StringTableBuilder, TableInitializer,
     },
     structures::{Map, TypedIndex},
     types::{
@@ -110,11 +110,36 @@ pub fn read_module<
     let mut namespace_builder = Some(StringTableBuilder::new());
     let mut name_builder = Some(StringTableBuilder::new());
     let mut func_count: usize = 0;
-    let mut mcg_info_fed = false;
+
+    let mut feed_mcg_signatures: Option<_> = Some(|mcg: &mut MCG| -> Result<(), LoadError> {
+        let info_read = info.read().unwrap();
+        mcg.feed_signatures(info_read.signatures.clone())
+            .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
+        Ok(())
+    });
+    let mut feed_mcg_info: Option<_> = Some(
+        |mcg: &mut MCG,
+         ns_builder: StringTableBuilder<NamespaceIndex>,
+         name_builder: StringTableBuilder<NameIndex>|
+         -> Result<(), LoadError> {
+            {
+                let mut info_write = info.write().unwrap();
+                info_write.namespace_table = ns_builder.finish();
+                info_write.name_table = name_builder.finish();
+            }
+            let info_read = info.read().unwrap();
+            mcg.feed_function_signatures(info_read.func_assoc.clone())
+                .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
+            mcg.check_precondition(&info_read)
+                .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
+            Ok(())
+        },
+    );
 
     loop {
         use wasmparser::ParserState;
         let state = parser.read();
+
         match *state {
             ParserState::Error(ref err) => return Err(err.clone().into()),
             ParserState::TypeSectionEntry(ref ty) => {
@@ -124,6 +149,10 @@ pub fn read_module<
                     .push(func_type_to_func_sig(ty)?);
             }
             ParserState::ImportSectionEntry { module, field, ty } => {
+                if let Some(f) = feed_mcg_signatures.take() {
+                    f(mcg)?;
+                }
+
                 let namespace_index = namespace_builder.as_mut().unwrap().register(module);
                 let name_index = name_builder.as_mut().unwrap().register(field);
                 let import_name = ImportName {
@@ -136,7 +165,7 @@ pub fn read_module<
                         let sigindex = SigIndex::new(sigindex as usize);
                         info.write().unwrap().imported_functions.push(import_name);
                         info.write().unwrap().func_assoc.push(sigindex);
-                        mcg.feed_import_function()
+                        mcg.feed_import_function(sigindex)
                             .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
                     }
                     ImportSectionEntryType::Table(table_ty) => {
@@ -217,23 +246,17 @@ pub fn read_module<
                 info.write().unwrap().start_func = Some(FuncIndex::new(start_index as usize));
             }
             ParserState::BeginFunctionBody { range } => {
-                let id = func_count;
-                if !mcg_info_fed {
-                    mcg_info_fed = true;
-                    {
-                        let mut info_write = info.write().unwrap();
-                        info_write.namespace_table = namespace_builder.take().unwrap().finish();
-                        info_write.name_table = name_builder.take().unwrap().finish();
-                    }
-                    let info_read = info.read().unwrap();
-                    mcg.feed_signatures(info_read.signatures.clone())
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
-                    mcg.feed_function_signatures(info_read.func_assoc.clone())
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
-                    mcg.check_precondition(&info_read)
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
+                if let Some(f) = feed_mcg_signatures.take() {
+                    f(mcg)?;
                 }
-
+                if let Some(f) = feed_mcg_info.take() {
+                    f(
+                        mcg,
+                        namespace_builder.take().unwrap(),
+                        name_builder.take().unwrap(),
+                    )?;
+                }
+                let id = func_count;
                 let fcg = mcg
                     .next_function(
                         Arc::clone(&info),
@@ -432,17 +455,15 @@ pub fn read_module<
                 info.write().unwrap().globals.push(global_init);
             }
             ParserState::EndWasm => {
-                // TODO Consolidate with BeginFunction body if possible
-                if !mcg_info_fed {
-                    info.write().unwrap().namespace_table =
-                        namespace_builder.take().unwrap().finish();
-                    info.write().unwrap().name_table = name_builder.take().unwrap().finish();
-                    mcg.feed_signatures(info.read().unwrap().signatures.clone())
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
-                    mcg.feed_function_signatures(info.read().unwrap().func_assoc.clone())
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
-                    mcg.check_precondition(&info.read().unwrap())
-                        .map_err(|x| LoadError::Codegen(format!("{:?}", x)))?;
+                if let Some(f) = feed_mcg_signatures.take() {
+                    f(mcg)?;
+                }
+                if let Some(f) = feed_mcg_info.take() {
+                    f(
+                        mcg,
+                        namespace_builder.take().unwrap(),
+                        name_builder.take().unwrap(),
+                    )?;
                 }
                 break;
             }